# Improve default security for SSLContext with a monkeypatch.