@jordansissel 2015-07-09
Other manual tests against `input {lumberjack { ... } }`

* TLSv1 w/ Export ciphers - correctly rejected:
```
% openssl s_client -tls1 -connect localhost:4444 -CAfile ~/projects/logstash-forwarder/lumberjack.crt -cipher "EXP"
CONNECTED(00000003)
6498:error:1409E0E5:SSL routines:SSL3_WRITE_BYTES:ssl handshake failure:/SourceCache/OpenSSL098/OpenSSL098-52.30.1/src/ssl/s3_pkt.c:566:
```

* SSLv3 with any ciphers: correctly rejected

```
% openssl s_client -ssl3 -connect localhost:4444 -CAfile ~/projects/logstash-forwarder/lumberjack.crt
CONNECTED(00000003)
6550:error:1409E0E5:SSL routines:SSL3_WRITE_BYTES:ssl handshake failure:/SourceCache/OpenSSL098/OpenSSL098-52.30.1/src/ssl/s3_pkt.c:566:
```

* SSLv2 with any ciphers: correctly rejected

```
% openssl s_client -ssl2 -connect localhost:4444 -CAfile ~/projects/logstash-forwarder/lumberjack.crt
CONNECTED(00000003)
6563:error:1407F0E5:SSL routines:SSL2_WRITE:ssl handshake failure:/SourceCache/OpenSSL098/OpenSSL098-52.30.1/src/ssl/s2_pkt.c:427:
```