@torrancew 2015-07-09
@jordansissel I would like to analyze this with something like sslyze as well. I've found it to be quite helpful for validating the behavior and selected cipher suites across a variety of platforms. I hope to do that tonight.