@jordansissel 2015-07-09
@torrancew good idea. I will wait for your analysis. I ran `sslyze --regular` against the lumberjack input just now and the results look OK. (Many `ClientCertificateRequested - Server requested a client certificate.` errors, but I take this to mean cipher handshake was OK and the server for some reason asked for a client cert)