@torrancew 2015-07-09
Indeed, --regular can emit some pretty weird options against a non-HTTP endpoint. When I get home tonight and have a chance to set up an LS using this patch, I'll run something like:
```sslyze --sslv2 --sslv3 --tlsv1 --tlsv1_1 --tlsv1_2 --reneg --hide_rejected_ciphers $host:$port```