@jsvd 2015-07-10
#### Test with nmap

Logstash config:
```
input {
  lumberjack {
    port => 443
    ssl_certificate => "/Users/joaoduarte/6805/server/server.crt"
    ssl_key => "/Users/joaoduarte/6805/server/server.key"
    type => "lumberjack"
    codec => "json"
  }
}
output {
   stdout { }
}
```


##### Before:

```
% nmap --script ssl-enum-ciphers node01

Starting Nmap 6.47 ( http://nmap.org ) at 2015-07-10 15:27 WEST
Nmap scan report for node01 (127.0.0.1)
Host is up (0.00085s latency).
rDNS record for 127.0.0.1: localhost
Not shown: 499 closed ports, 499 filtered ports
PORT    STATE SERVICE
443/tcp open  https
| ssl-enum-ciphers: 
|   SSLv3: 
|     ciphers: 
|       TLS_DHE_RSA_EXPORT_WITH_DES40_CBC_SHA - weak
|       TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA - strong
|       TLS_DHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_DHE_RSA_WITH_DES_CBC_SHA - weak
|       TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_RSA_EXPORT_WITH_DES40_CBC_SHA - weak
|       TLS_RSA_EXPORT_WITH_RC4_40_MD5 - weak
|       TLS_RSA_WITH_3DES_EDE_CBC_SHA - strong
|       TLS_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_RSA_WITH_DES_CBC_SHA - weak
|       TLS_RSA_WITH_RC4_128_MD5 - strong
|       TLS_RSA_WITH_RC4_128_SHA - strong
|     compressors: 
|       NULL
|   TLSv1.0: 
|     ciphers: 
|       TLS_DHE_RSA_EXPORT_WITH_DES40_CBC_SHA - weak
|       TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA - strong
|       TLS_DHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_DHE_RSA_WITH_DES_CBC_SHA - weak
|       TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_RSA_EXPORT_WITH_DES40_CBC_SHA - weak
|       TLS_RSA_EXPORT_WITH_RC4_40_MD5 - weak
|       TLS_RSA_WITH_3DES_EDE_CBC_SHA - strong
|       TLS_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_RSA_WITH_DES_CBC_SHA - weak
|       TLS_RSA_WITH_RC4_128_MD5 - strong
|       TLS_RSA_WITH_RC4_128_SHA - strong
|     compressors: 
|       NULL
|_  least strength: weak
631/tcp open  ipp

Nmap done: 1 IP address (1 host up) scanned in 3.55 seconds
```

##### After:

```
% nmap --script ssl-enum-ciphers node01       

Starting Nmap 6.47 ( http://nmap.org ) at 2015-07-10 15:26 WEST
Nmap scan report for node01 (127.0.0.1)
Host is up (0.00014s latency).
rDNS record for 127.0.0.1: localhost
Not shown: 998 closed ports
PORT    STATE SERVICE
443/tcp open  https
| ssl-enum-ciphers: 
|   TLSv1.0: 
|     ciphers: 
|       TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA - strong
|       TLS_DHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA - strong
|       TLS_RSA_WITH_AES_128_CBC_SHA - strong
|     compressors: 
|       NULL
|_  least strength: strong
631/tcp open  ipp

Nmap done: 1 IP address (1 host up) scanned in 6.41 seconds
```

It's strange that TLSv1.1 and 2 aren't supported in either case, maybe something to do with the certificate  I used?